fix: extract RomeDetail et InfosDiffusionOffre
