+++ ui/components/espace_pro/AjouterVoeux.tsx
@@ -1,8 +1,4 @@
 import {
-  Accordion,
-  AccordionButton,
-  AccordionItem,
-  AccordionPanel,
   Box,
   Button,
   Checkbox,
@@ -14,7 +10,6 @@ import {
   FormHelperText,
   FormLabel,
   Heading,
-  Image,
   Input,
   Link,
   Select,
@@ -41,9 +36,10 @@ import { debounce } from "../../common/utils/debounce"
 import { publicConfig } from "../../config.public"
 import { LogoContext } from "../../context/contextLogo"
 import { WidgetContext } from "../../context/contextWidget"
-import { ArrowRightLine, ExternalLinkLine, InfoCircle, Minus, Plus, Warning } from "../../theme/components/icons"
-import { J1S, Parcoursup } from "../../theme/components/logos_pro"
+import { ArrowRightLine, ExternalLinkLine, Minus, Plus, Warning } from "../../theme/components/icons"
 import { createOffre, createOffreByToken, getFormulaire, getFormulaireByToken, getRelatedEtablissementsFromRome, getRomeDetail } from "../../utils/api"
+import { InfosDiffusionOffre } from "../DepotOffre/InfosDiffusionOffre"
+import { RomeDetail } from "../DepotOffre/RomeDetail"
 
 import DropdownCombobox from "./DropdownCombobox"
 
@@ -194,7 +190,6 @@ const AjouterVoeuxForm = (props) => {
         job_type: props.job_type ?? ["Apprentissage"],
         is_multi_published: props.is_multi_published ?? undefined,
         delegations: props.delegations ?? undefined,
-        rome_detail: props.rome_detail ?? {},
         is_disabled_elligible: props.is_disabled_elligible ?? false,
         job_count: props.job_count ?? 1,
         job_duration: props.job_duration ?? 12,
@@ -231,7 +226,7 @@ const AjouterVoeuxForm = (props) => {
                     setFieldValue("rome_code", [values.codeRome])
                   }, 0)
 
-                  props.getRomeInformation(values.codeRome, values.appellation, formik)
+                  props.getRomeInformation(values.codeRome, values.appellation)
                 }}
                 name="rome_label"
                 value={values.rome_appellation_label}
@@ -374,134 +369,18 @@ const AjouterVoeuxForm = (props) => {
   )
 }
 
-const RomeInformationDetail = ({ definition, competencesDeBase, libelle, appellation, acces }) => {
-  if (definition) {
-    const definitionSplitted = definition.split("\\n")
-    const accesFormatted = acces.split("\\n").join("<br><br>")
-
-    return (
-      <Box border="1px solid #000091" p={5} mb={5}>
-        <Box mb={5}>
-          <Heading fontSize="24px" mb={3}>
-            {appellation}
-          </Heading>
-          <Text fontSize="16px" fontWeight="700">
-            Fiche métier : {libelle}
-          </Text>
-          <Text fontSize="14px">La fiche métier se base sur la classification ROME de France Travail</Text>
-        </Box>
-        <Flex alignItems="flex-start" mb={6}>
-          <InfoCircle mr={2} mt={1} color="bluefrance.500" />
-          <Text textAlign="justify">Voici la description visible par les candidats lors de la mise en ligne de l’offre d’emploi en alternance.</Text>
-        </Flex>
-
-        <Accordion defaultIndex={[0]} allowMultiple>
-          <AccordionItem key={0} id="metier">
-            {({ isExpanded }) => (
-              <>
-                <h2>
-                  <AccordionButton>
-                    <Text fontWeight="700" flex="1" textAlign="left">
-                      Description du métier
-                    </Text>
-                    {isExpanded ? <Minus color="bluefrance.500" /> : <Plus color="bluefrance.500" />}
-                  </AccordionButton>
-                </h2>
-                <AccordionPanel pb={4} ml={6} mr={3}>
-                  <ul className="voeuxUl">
-                    {definitionSplitted.map((x) => {
-                      return (
-                        <li className="voeuxUlLi" key={x}>
-                          {x}
-                        </li>
-                      )
-                    })}
-                  </ul>
-                </AccordionPanel>
-              </>
-            )}
-          </AccordionItem>
-          <hr />
-          <AccordionItem key={1} id="competence">
-            {({ isExpanded }) => (
-              <>
-                <h2>
-                  <AccordionButton>
-                    <Text fontWeight="700" flex="1" textAlign="left">
-                      Quelles sont les compétences visées ?
-                    </Text>
-                    {isExpanded ? <Minus color="bluefrance.500" /> : <Plus color="bluefrance.500" />}
-                  </AccordionButton>
-                </h2>
-                <AccordionPanel maxH="50%" pb={4} ml={6} mr={3}>
-                  <ul className="voeuxUl">
-                    {competencesDeBase.map((x) => (
-                      <li className="voeuxUlLi" key={x.libelle}>
-                        {x.libelle}
-                      </li>
-                    ))}
-                  </ul>
-                </AccordionPanel>
-              </>
-            )}
-          </AccordionItem>
-          <hr />
-          <AccordionItem key={2} id="accessibilite">
-            {({ isExpanded }) => (
-              <>
-                <h2>
-                  <AccordionButton>
-                    <Text fontWeight="700" flex="1" textAlign="left">
-                      À qui ce métier est-il accessible ?
-                    </Text>
-                    {isExpanded ? <Minus color="bluefrance.500" /> : <Plus color="bluefrance.500" />}
-                  </AccordionButton>
-                </h2>
-                <AccordionPanel maxH="50%" pb={4}>
-                  <span dangerouslySetInnerHTML={{ __html: accesFormatted }}></span>
-                </AccordionPanel>
-              </>
-            )}
-          </AccordionItem>
-        </Accordion>
-      </Box>
-    )
-  } else {
-    return (
-      <Box border="1px solid #000091" p={5}>
-        <Heading fontSize="24px" mb={3}>
-          Dites-nous en plus sur votre besoin en recrutement
-        </Heading>
-        <Flex alignItems="flex-start" mb={6} justify="flex-start">
-          <InfoCircle mr={2} mt={1} color="bluefrance.500" />
-          <Text textAlign="justify">Cela permettra à votre offre d'être visible des candidats intéressés.</Text>
-        </Flex>
-        <Box ml={5}>
-          <Text>Une fois créée, votre offre d’emploi sera immédiatement mise en ligne sur les sites suivants : </Text>
-          <Stack direction={["column", "row"]} spacing={2} align="center" justify="center" mt={3}>
-            <Image src="/images/logo_LBA.svg" alt="" minWidth="150px" width="150px" />
-            <J1S w="100px" h="100px" />
-            <Parcoursup w="220px" h="100px" />
-          </Stack>
-        </Box>
-      </Box>
-    )
-  }
-}
-
 export const PageAjouterVoeux = (props) => {
-  const [romeInformation, setRomeInformation] = useState({})
+  const [romeInformation, setRomeInformation] = useState(null)
   const [loading, setLoading] = useState(false)
   const { widget } = useContext(WidgetContext)
 
   const router = useRouter()
   const { establishment_id } = router.query
 
-  const getRomeInformation = (rome: string, appellation, formik) => {
+  const getRomeInformation = (rome: string, appellation) => {
     getRomeDetail(rome)
       .then((data) => {
         setLoading(true)
-        formik.setFieldValue("rome_detail", data)
         setRomeInformation({ appellation, ...data })
       })
       .catch((error) => console.error(error))
@@ -536,8 +415,10 @@ export const PageAjouterVoeux = (props) => {
             <Spinner thickness="4px" speed="0.5s" emptyColor="gray.200" color="bluefrance.500" size="xl" />
             <Text>Recherche en cours...</Text>
           </Flex>
+        ) : romeInformation ? (
+          <RomeDetail {...(romeInformation as any)} />
         ) : (
-          <RomeInformationDetail {...(romeInformation as any)} />
+          <InfosDiffusionOffre />
         )}
       </Box>
     </SimpleGrid>
